+++ src/app/dialogs/dialog-large-format-page/dialog-large-format-page.module.ts
@@ -1,0 +1,23 @@
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { DialogModule, DrawerModule, SectionModule, StepperModule, TabsModule } from '@swimlane/ngx-ui';
+import { PrismModule } from '../../common/prism/prism.module';
+import { ChildDrawerComponent } from './components/child-drawer.component';
+import { ParentDrawerComponent } from './components/parent-drawer.component';
+import { DialogLargeFormatDialogPageComponent } from './dialog-large-format-dialog-page.component';
+import { DialogLargeFormatPageRoutingModule } from './dialog-large-format-page-routing.module';
+
+@NgModule({
+  declarations: [DialogLargeFormatDialogPageComponent, ParentDrawerComponent, ChildDrawerComponent],
+  imports: [
+    CommonModule,
+    DialogLargeFormatPageRoutingModule,
+    StepperModule,
+    DrawerModule,
+    TabsModule,
+    DialogModule,
+    SectionModule,
+    PrismModule
+  ]
+})
+export class DialogLargeFormatPageModule {}