+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Feature(CardComponent): add error outline
 - Fix(DropzoneComponent): addressing minor design review feedback
+- Feat(LargeFormatDialogContentComponent): add new component as wrapper for Dialog content
 
 =======